ui: add KeyHandler class and modify more code
The key event is handled by ui.KeyHandler class.  keyhandler.getkey()
returns a more lucid character or string.  Therefore, the method of
receiving the key event must change the number of arguments.  Moreover,
it becomes easy to insert in the command line.  Because the input of
multi byte character in utf8 does buffering.  The keymap module became
unnecessary by this change.  Therefore, it is deleted.  Unfortunately,
rear compatibility of rc.py is greatly lost.