completion: add Candidate class for completion candidate
A Candidate class is a subclass that inherits to a listbox.Entry class
for completion candidate.  Especially, this class has plural names and
documents.  This attributes can add plural names and explanations to
completion candidates.  Also, a Argument class that inherits to Candidate
class has a callback function of completion candidate which deals with
oneself.  Therefore, this implementation is useful for completion with
a option of program.